For substantive comments see full commit message
Responsibility lies with a person, and we already have identified (in
reality) who that person is - the Documentation Engineer (@jsick at this
time). I don't object to strongly to abstracting it to a TechDocs team
terminology if that makes it map industry conecpts (eg. the Twitter
reference) if that brings useful semantics to the conversation, but
ultimately any effort expended here is simply a scaling of the
Documentation Engineer.