Substantive comment in the long commit message.
In general I am very uncomfortable with suggesting that we accept Google
Docs and Wiki pages as primary documentation formats. I can think of a
number of reasons where this is a sub optimal. I think a case should
need to be made for the workflows that are described here to apply. Or
really, we could be Nazis and just say DocHub may provide discovery but
we do not curate the content in LSST-The-Docs (frankly I am leaning that
way and would like to hear arguments as to the contrary).